Latch i_w instead of and-ing the clock in shift registers
And-ing the clock signal resulted in the shift register detecting
an additional clock transition, shifting 1 additional time.
Apparently having multiple clock signals is considered bad practice,
I found out the hard way.

Instead, use an enable signal and latch i_w, so that it gets delayed 1
cycle and we can change the enable signal in 1 cycle and read the next.